requirements.txt
alembic==1.8.1
anyio==3.6.1
asgiref==3.5.2
attrs==22.1.0
beautifulsoup4==4.11.1
blis==0.7.8
breadability==0.1.20
bs4==0.0.1
catalogue==2.0.8
certifi==2022.6.15
chardet==5.0.0
charset-normalizer==2.1.0
click==8.1.3
cymem==2.0.6
de-core-news-md @ https://github.com/explosion/spacy-models/releases/download/de_core_news_md-3.3.0/de_core_news_md-3.3.0.tar.gz
docopt==0.6.2
en-core-web-md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.3.0/en_core_web_md-3.3.0.tar.gz
es-core-news-md @ https://github.com/explosion/spacy-models/releases/download/es_core_news_md-3.3.0/es_core_news_md-3.3.0.tar.gz
et-xmlfile==1.1.0
fastapi==0.78.0
FastAPI-SQLAlchemy==0.2.1
fastapi-versioning==0.10.0
fr-core-news-md @ https://github.com/explosion/spacy-models/releases/download/fr_core_news_md-3.3.0/fr_core_news_md-3.3.0.tar.gz
greenlet==1.1.2
h11==0.13.0
httptools==0.4.0
idna==3.3
iniconfig==1.1.1
it-core-news-md @ https://github.com/explosion/spacy-models/releases/download/it_core_news_md-3.3.0/it_core_news_md-3.3.0.tar.gz
Jinja2==3.1.2
joblib==1.1.0
langcodes==3.3.0
langdetect==1.0.9
lxml==4.9.1
Mako==1.2.1
MarkupSafe==2.1.1
murmurhash==1.0.7
nltk==3.7
numpy==1.23.1
openpyxl==3.0.10
packaging==21.3
pandas==1.4.2
pathy==0.6.2
pluggy==1.0.0
preshed==3.0.6
psycopg2==2.9.3
pt-core-news-md @ https://github.com/explosion/spacy-models/releases/download/pt_core_news_md-3.3.0/pt_core_news_md-3.3.0.tar.gz
py==1.11.0
pycountry==22.3.5
pydantic==1.8.2
PyJWT==2.4.0
pyparsing==3.0.9
pytest==7.1.2
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
PyYAML==6.0
regex==2022.7.25
requests==2.28.1
six==1.16.0
smart-open==5.2.1
sniffio==1.2.0
soupsieve==2.3.2.post1
spacy==3.3.0
spacy-legacy==3.0.9
spacy-loggers==1.0.3
SQLAlchemy==1.4.40
srsly==2.4.4
starlette==0.19.1
sumy==0.10.0
thinc==8.0.17
tokenizers==0.12.1
tomli==2.0.1
tqdm==4.64.0
typer==0.4.2
typing_extensions==4.3.0
urllib3==1.26.11
uvicorn==0.17.6
uvloop==0.16.0
wasabi==0.10.1
watchgod==0.8.2
websockets==10.3
xmltodict==0.13.0
